svcgssd: use correct defaults in call to gss_set_allowable_enctypes
For the window of kernels between 2.6.35 (when the support
for newer encryption was added) and 2.6.39 (when the ability
to read the supported enctypes from the kernel was added),
use a default of all enctypes when the kernel supported
enctypes file cannot be read.

For kernels before 2.6.35, continue to use a default of
only DES enctypes.

Note that the version of Kerberos must also support the use of
gss_set_allowable_enctypes for service-side negotiations.

See also: http://bugzilla.redhat.com/show_bug.cgi?id=719776

Signed-off-by: Kevin Coffman <kwc@citi.umich.edu>
Signed-off-by: Steve Dickson <steved@redhat.com>