Fix LSB init script output
For the exporting/unexporting action use log_action_*_msg instead
of log_*_msg, so that we switch from this output:

  start → Exporting directories for NFS kernel daemon....
  stop → Unexporting directories for NFS kernel daemon....

to:

  start → Exporting directories for NFS kernel daemon...done.
  stop → Unexporting directories for NFS kernel daemon...done.

which is confusing, given that there are three dots for the initial
action plus the final dot for completion.

Remove trailing dot from log_warning_msg calls, to switch from stuff
like:

  Not starting NFS kernel daemon: no exports. ... (warning).

to:

  Not starting NFS kernel daemon: no exports ... (warning).

where the trailing dot is not misplaced anymore.

Use status_of_proc instead of an ad-hoc implementation.