Propagate $RPCGSSDOPTS from /etc/default/nfs-utils to systemd service script. (Closes: #846950)
Pursuant to Salvatore's request [1] this fixes, in a minimal way, a
naming inconsistency that keeps command line arguments for rpc.gssd
from being propagated between

    /etc/default/nfs-common

and

    /lib/systemd/system/rpc-gssd.service

Users of separate keytabs in /etc/keytabs still have to make other
manual changes, but this fixes the error that is hardest to track
down. It is a typo, and has remained unresolved for over for years!

Whenever nfs-utils is upgraded, I fix it manually on all my systems.
As someone who filed or amended the involved bugs years ago
ago, I disagreed with the administrative downgrade by Gaudenz
Steinlin. Raising them again was on my todo list for bullseye.

Please allow at least this small correction of a typo into the
upcoming release. Thank you!

Eventually, it would be better to integrate Joachim Falk's merge
request [2] into the changes are pending from upstream. I did not yet
test it, but he seems to address the failing conditions that cause
systemd to quit due to some paths not existing in my use case.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=846950#84
[2] https://salsa.debian.org/kernel-team/nfs-utils/-/merge_requests/5

[carnil: Make changelog entry UNRELEASED]