Stop shipping /var/lib/nfs/{etab,rmtab,state} (Closes: #1074359)
We currently ship empty files /var/lib/nfs/{etab,rmtab} in
nfs-kernel-server and /var/lib/nfs/state in nfs-common, as well as
creating and removing some of them in maintainer scripts.  But these
are local state files that should not be overwritten in upgrades.

As a first step, stop shipping them.  This will result in later
upgrades doing the right thing, but will cause them to be removed
during an upgrade from an older version.

In nfs-common postinst, create /var/lib/nfs/state if missing.
nfs-kernel-server already created its state files if missing.