d/control: nfs-common: Move python3 to Recommends
python3 is used in the mountstats, nfsiostat and rpcctl tool shipped in
the nfs-common package. While python3 is required to run these tools,
for the nfs-common package as whole it is a strong, but not absolute
necessary dependency and we can avoid introducing another binary package
just to ship those three tools.

Making python3 a Recommends should help minimal installations which do
not need or want to pull in python3.

Closes: #1013868