Move nfs-blkmap.service to nfs-common package
The blkmapd daemon is only required on NFS clients which are using pNFS.
In particular we did already install the blkmap tool into the nfs-common
package but not the corresponding service.

Move the nfs-blkmap.service from nfs-kernel-server package to
nfs-common.

Add Replaces and Breaks relation for nfs-common package on the
nfs-kernel-server package with (<< 1:2.8.2-2~exp1~) constraints to
properly take over the nfs-blkmap.service service file.

Closes: #1072611